Create external chat gpi call function
